+++ Sources/JWTKeychain/Models/JWTKeychainUser/Extensions/JWTKeychainUser+JWTKeychainUserType.swift
@@ -1,0 +1,1 @@
+extension JWTKeychainUser: JWTKeychainUserType {}
