+++ Sources/JWTKeychain/Models/JWTKeychainUser/Extensions/JWTKeychainUser+PasswordAuthenticatable.swift
@@ -1,0 +1,10 @@
+import Authentication
+
+extension JWTKeychainUser: PasswordAuthenticatable {
+
+    /// See 'BasicAuthenticatable'.
+    public static let passwordKey: WritableKeyPath<JWTKeychainUser, String> = \.password
+
+    /// See 'BasicAuthenticatable'.
+    public static let usernameKey: WritableKeyPath<JWTKeychainUser, String> = \.email
+}
